requirements.txt
flask
flask-socketio
spotipy
python-dotenv

# Pour déploiement en prod (optionnel)
gunicorn
eventlet